Adding multi-search/delimited search
Using the | character, users can search the list of attendees for multiple users at the same time. Particularly useful when the list of users is huge, to avoid needing to enrol users one at a time.

Recommendation: Works well with the Excel TEXTJOIN() function

Contributions by @chalhoub-university, @geoffreyvanwyk and me